+++ Tests/SourceKittenFrameworkTests/SwiftDocsTests.swift
@@ -67,7 +67,9 @@ private func compareDocs(withFixtureNamed name: String, file: StaticString = #fi
 }
 
 private func versionedExpectedFilename(for name: String) -> String {
-#if compiler(>=5.3)
+#if compiler(>=5.3.1)
+    let versions = ["swift-5.3.1", "swift-5.3", "swift-5.2", "swift-5.1", "swift-5.0"]
+#elseif compiler(>=5.3)
     let versions = ["swift-5.3", "swift-5.2", "swift-5.1", "swift-5.0"]
 #elseif compiler(>=5.2)
     let versions = ["swift-5.2", "swift-5.1", "swift-5.0"]
@@ -108,7 +110,11 @@ private func diff(original: String, modified: String) -> String {
 }
 
 private let buildingSwiftVersion: String = {
-#if compiler(>=5.2)
+#if compiler(>=5.3.1)
+    return "swift-5.3.1"
+#elseif compiler(>=5.3)
+    return "swift-5.3"
+#elseif compiler(>=5.2)
     return "swift-5.2"
 #elseif compiler(>=5.1)
     return "swift-5.1"